// front-end/src/app/models/rddv/rddv.model.ts
import { RddvAttachmentModel } from "./rddv-attachment.model";
import { RddvExpenseModel } from "./rddv-expense.model";

export class RddvModel {
    idRelatorio: number;
    idUsuario: number;
    nomeFuncionario: string;
    idUsuarioAtribuido: number | null;
    dataAtribuicao: Date | null;
    idUsuarioAtribuidor: number | null;
    cpf: string;
    idDepartamento: number;
    finalidade: string;
    tipoRelatorio: number;
    moeda: string;
    tipoViagem: number;
    dataInicio: Date | null;
    dataFim: Date | null;
    dataCadastro: Date;
    dataVencimento: Date | null;
    localViagem: string;
    centroCusto: string;
    projeto: string;
    pca: string;
    diariaViagem: boolean | null;
    diarias: number | null;
    adiantamento: boolean | null;
    valorAdiantamento: number;
    saldoCartaoVtm: number;
    observacao: string;
    tipoAutorizacao: number;
    idGestor: number | null;
    nomeGestor: string | null;
    aprovadoGestor: boolean | null;
    observacaoGestor: string;
    aprovadoSetor: boolean | null;
    observacaoSetor: string;
    statusGestor: number | null;
    statusContabilidade: number | null;
    statusFinanceiro: number | null;
    despesas: RddvExpenseModel[];
    documentos: RddvAttachmentModel[];
    valorTotal: number;
    banco: string;
    agencia: string;
    conta: string;
    valorDiaria: number;
    valorKm: number;
    dataAprovacaoGestor: string | null;
    dataAprovacaoFinanceiro: string | null;
    dataAprovacaoContabil: string | null;
    rascunho: boolean = true;
    cancelado: boolean = false;
}